% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_perf_infer.R
\name{boot_infer}
\alias{boot_infer}
\title{Bootstrap Inference}
\usage{
boot_infer(rets, boot_reps = 100, blocks_n = 1, type = "var", digits = 5)
}
\arguments{
\item{rets}{an nx2 matrix with two returns time series.
n stands for the observation length.
Only two return time series can be compared simultaneously.}

\item{boot_reps}{an integer, the number of bootstrap repetitions.}

\item{blocks_n}{an integer, the number of blocks for the circular block bootstrap.
Default value is 1.}

\item{type}{a character, indicating the performance measure (parameter) of interest.
Possible values are type=c("var", "sharpe"). Default value is type="var".}

\item{digits}{an integer, indicating how many digits the results are to be rounded to. Default value is 5.}
}
\value{
a list
\itemize{
\item param-diff, the difference between the estimated parameters in type
(first minus second time series).
For type="var", the function returns the differences in log-variances.
\item se, the estimated standard error of the parameter in type.
\item p-value, the estimated p-value.
}
}
\description{
Performs a studentized Bootstrap inference
}
\examples{
data(prices_m)
rets_m <- calc_rets(prices_m)
sigma_mat <- cov(rets_m)
mu_vec <- mu_estim_ml(rets_m)
port_naive <- port_optim_naive(sigma_mat)
port_tang <- port_optim_tang(sigma_mat, mu_vec)
port_rets <- cbind(calc_port_rets(port_naive, rets_m), calc_port_rets(port_tang, rets_m))
boot_infer_res <- boot_infer(port_rets)

}
