% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_perf_infer.R
\name{boot_nonparam}
\alias{boot_nonparam}
\title{Nonparametric Bootstrap}
\usage{
boot_nonparam(rets, boot_reps, sample_size, type = calc_sr)
}
\arguments{
\item{rets}{an nxp stock returns matrix.}

\item{boot_reps}{an integer, the number of bootstrap repetitions.}

\item{sample_size}{an integer, the sample size for the nonparametric bootstrap.}

\item{type}{a function, the performance measure to be calculated. Default is calc_sr.}
}
\value{
a Bxp data matrix with the bootstrapped values of the performance measure from type.
}
\description{
Performs a nonparametric bootstrap on returns time series and a predefined performance measure.
}
\examples{
data(prices_m)
rets_m <- calc_rets(prices_m)
srs_boot <- boot_nonparam(rets_m, boot_reps = 10, sample_size = 50, type = calc_sr)

}
