% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_perf_infer.R
\name{boot_param}
\alias{boot_param}
\title{Parametric Bootstrap}
\usage{
boot_param(rets, boot_reps, sample_size, type = calc_sr)
}
\arguments{
\item{rets}{an nxp stock returns matrix.}

\item{boot_reps}{an integer, the number of bootstrap repetitions.}

\item{sample_size}{an integer, the sample size for the parametric bootstrap.}

\item{type}{a function, the performance measure to be calculated. Default is sr_calc.}
}
\value{
a Bxp data matrix with the bootstrapped values of the performance measure from type.
}
\description{
Performs a parametric bootstrap on returns time series (with an assumed normal distribution)
and predefined performance measure.
}
\examples{
data(prices_m)
rets_m <- calc_rets(prices_m)
srs_boot <- boot_param(rets_m, boot_reps = 10, sample_size = 50, type = calc_sr)

}
