% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_perf_rets.R
\name{calc_mean}
\alias{calc_mean}
\title{Expected Return of Returns}
\usage{
calc_mean(rets, ann_factor = 1)
}
\arguments{
\item{rets}{a numerical vector of stock or portfolio returns.}

\item{ann_factor}{a double, the annualization factor. If ann_factor=1 (default), no annualization is performed.
For monthly returns, set ann_factor=12. For daily returns, set ann_factor=252, etc.}
}
\value{
a double the expected return.
}
\description{
Calculates the expected return of stock returns.
}
\examples{
data(prices_spx_m)
rets_spx_m <- calc_rets(prices_spx_m)
spx_mean <- calc_mean(rets_spx_m)
data(prices_m)
rets_m <- calc_rets(prices_m)
stocks_mean <- apply(rets_m, 2, calc_mean)

}
