% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_perf_rets.R
\name{calc_port_mu_is}
\alias{calc_port_mu_is}
\title{Portfolio Expected Returns (in-sample)}
\usage{
calc_port_mu_is(port_weights, mu_vec)
}
\arguments{
\item{port_weights}{a numeric vector, the portfolio weights.}

\item{mu_vec}{a numeric vector of expected returns.}
}
\value{
a double, the portfolio expected return.
}
\description{
Calculates the in-sample portfolio expected returns.
}
\examples{
data(prices_m)
rets_m <- calc_rets(prices_m)
sigma_mat <- cov(rets_m)
mu_vec <- mu_estim_ml(rets_m)
port_naive <- port_optim_naive(sigma_mat)
port_mu_is <- calc_port_mu_is(port_naive, mu_vec)

}
