% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_perf_turn.R
\name{calc_port_turnover}
\alias{calc_port_turnover}
\title{Portfolio Turnover}
\usage{
calc_port_turnover(port_weights)
}
\arguments{
\item{port_weights}{a numerical nxp data matrix with portfolio weights over the whole observation period.}
}
\value{
a numeric vector with the turnover over the specified observation period.
}
\description{
Calculates the average turnover of a portfolio strategy.
}
\examples{
set.seed(123)
# Simulated weights matrix (each row is a time period, columns are assets)
port_sim <- matrix(runif(100, min = 0, max = 1), nrow = 10, ncol = 10)
port_sim <- port_sim / rowSums(port_sim)
port_turn <- calc_port_turnover(port_sim)

}
