% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_perf_rets.R
\name{calc_port_var_is}
\alias{calc_port_var_is}
\title{Portfolio Variance (in-sample)}
\usage{
calc_port_var_is(port_weights, sigma_mat)
}
\arguments{
\item{port_weights}{a numeric vector, the portfolio weights.}

\item{sigma_mat}{a pxp covariance matrix of returns.}
}
\value{
a double, the portfolio variance.
}
\description{
Calculates the in-sample portfolio variance.
}
\examples{
data(prices_m)
rets_m <- calc_rets(prices_m)
sigma_mat <- cov(rets_m)
port_naive <- port_optim_naive(sigma_mat)
port_var_is <- calc_port_var_is(port_naive, sigma_mat)

}
