% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_help_funcs.R
\name{calc_rets}
\alias{calc_rets}
\title{Calculate Stock Returns}
\usage{
calc_rets(prices, type = "log")
}
\arguments{
\item{prices}{a time series object representing the prices.}

\item{type}{a character, identifying the type of returns. Possible values are "log" for log returns
and "simple" for simple returns. Default value is set to "log".}
}
\value{
a time series object with log returns.
}
\description{
This function computes the stock returns of a given vector or time series of prices.
Log returns are calculated as the natural logarithm of the ratio of consecutive prices.
}
\examples{
data(prices_m)
rets_m <- calc_rets(prices_m)
head(rets_m[, 1:5])
rets_spx_m <- calc_rets(prices_spx_m)
head(rets_spx_m)
}
