% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_perf_rets.R
\name{calc_sr}
\alias{calc_sr}
\title{Sharpe Ratio of Returns}
\usage{
calc_sr(rets, rf = 0, ann_factor = 1)
}
\arguments{
\item{rets}{a numerical vector of stock or portfolio returns.}

\item{rf}{a double, the assumed risk-free return. Default=0.}

\item{ann_factor}{a double, the annualization factor. If ann_factor=1 (default), no annualization is performed.
For monthly returns, set ann_factor=12. For daily returns, set ann_factor=252, etc.}
}
\value{
a double the Sharpe ratio.
}
\description{
Calculates the Sharpe ratio of returns.
}
\examples{
data(prices_spx_m)
rets_spx_m <- calc_rets(prices_spx_m)
spx_sharpe <- calc_sr(rets_spx_m)
data(prices_m)
rets_m <- calc_rets(prices_m)
stocks_sharpe <- apply(rets_m, 2, calc_sr)

}
