% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_perf_rets.R
\name{calc_var}
\alias{calc_var}
\title{Variance of Returns}
\usage{
calc_var(rets, ann_factor = 1)
}
\arguments{
\item{rets}{a numerical vector of stock or portfolio returns.}

\item{ann_factor}{a double, the annualization factor. If ann_factor=1 (default), no annualization is performed.
For monthly returns, set ann_factor=12. For daily returns, set ann_factor=252, etc.}
}
\value{
a double the variance.
}
\description{
Calculates the variance of stock returns.
}
\examples{
data(prices_spx_m)
rets_spx_m <- calc_rets(prices_spx_m)
spx_var <- calc_var(rets_spx_m)
data(prices_m)
rets_m <- calc_rets(prices_m)
stocks_var <- apply(rets_m, 2, calc_var)

}
