% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_perf_infer.R
\name{f_test}
\alias{f_test}
\title{F-Test
Performs a two-sided F-test to compare the variances of two samples from normal populations.}
\usage{
f_test(rets)
}
\arguments{
\item{rets}{an nx2 matrix with two returns time series.
n stands for the observation length.
Only two return time series can be compared simultaneously.}
}
\value{
a list
\itemize{
\item vars, the estimated variances.
\item vars-ratio, the ratio of the estimated variances.
\item p-value, the estimated p-value.
}
}
\description{
F-Test
Performs a two-sided F-test to compare the variances of two samples from normal populations.
}
\details{
The null hypothesis is that the ratio of the variances of the populations
from which the returns were drawn is equal to 1,
that is, the variances are equal.
The F-test requires that the data come from a bivariate normal distribution
without correlation and dependence over time.
If the data are correlated, have tails heavier than the normal distribution, or are dependent over time,
the test is not valid \insertCite{LedoitWolf_2011}{portoptim}.
}
\examples{
data(prices_m)
rets_m <- calc_rets(prices_m)
sigma_mat <- cov(rets_m)
mu_vec <- mu_estim_ml(rets_m)
port_naive <- port_optim_naive(sigma_mat)
port_tang <- port_optim_tang(sigma_mat, mu_vec)
port_rets <- cbind(calc_port_rets(port_naive, rets_m), calc_port_rets(port_tang, rets_m))
ftest_res <- f_test(port_rets)

}
