% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ff_5factors.R
\docType{data}
\name{ff_5factors}
\alias{ff_5factors}
\title{5 Fama-French Research Factors}
\format{
A data frame (216x6) with the following columns:
\describe{
\item{Mkt.RF}{Market - Risk Free}
\item{SMB}{Small Minus Big}
\item{HML}{High Minus Low}
\item{RMW}{Robust Minus Weak}
\item{CMA}{Conservative Minus Aggressive}
\item{RF}{Risk Free}
}
}
\usage{
data(ff_5factors)
}
\description{
This dataset contains the monthly returns for the five Fama-French Research Factors (MRP, SMB, HML, RMW, CMA)
for the time period from 01.2001 to 12.2018. The dataset was created with the package 'ffdata'.
}
\keyword{datasets}
