% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_help_funcs.R
\name{find_rep_vals}
\alias{find_rep_vals}
\title{Find Repeated Values in a Data Frame}
\usage{
find_rep_vals(df, max_rep = 10)
}
\arguments{
\item{df}{a data frame in which to search for repeated values.}

\item{max_rep}{an integer, specifying the maximum number of repetitions to consider.
Defaults to 10.}
}
\value{
a summary of the positions of repeated values in the data frame, up to the specified maximum.
}
\description{
This function identifies repeated values in a data frame and returns a summary
of the repetitions, up to a specified maximum number of repetitions.
}
\examples{
df <- data.frame(a = c(1, 2, 2, 3, 3, 3), b = c(4, 5, 5, 6, 6, 6))
find_rep_vals(df, max_rep = 2)

}
