% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_mu_estim.R
\name{mu_estim_capm}
\alias{mu_estim_capm}
\title{Single-Factor Expected Returns}
\usage{
mu_estim_capm(rets, market_rets_excess, rf = 0)
}
\arguments{
\item{rets}{an nxp matrix, the stock returns. n stands for the observation length, p for the number of stocks.}

\item{market_rets_excess}{an nx1 vector, the excess market return.}

\item{rf}{a risk-free rate of return. Default value is rf=0.}
}
\value{
a numeric vector with the estimated expected returns.
}
\description{
Estimates the expected returns with a single-factor model, specifically the Capital Asset Pricing Model (CAPM).
}
\details{
The Single-Factor estimator for each expected return \eqn{i} of an \eqn{n\times p} matrix
with stock returns \eqn{R} is
calculated with the following formula:
\deqn{E[R_{i}]=r_{f} + \widehat{\beta}_{i}\left(E[R_{M}]-r_{f}\right),}
where \eqn{r_{f}} is the risk-free rate of return, \eqn{E[R_{M}]} is the expected market return and
\eqn{\widehat{{\beta}}_{i}} are the estimated beta coefficients from the linear regression
\deqn{R_{ij} - r_{f} = \alpha_{i} + \beta_{i}\left(R_{Mj}-r_{f}\right) + \epsilon_{ij}}
for \eqn{i=1\cdots p} and \eqn{j=1\cdots n}.
}
\examples{
data(prices_m)
data(ff_5factors)
rets_m <- calc_rets(prices_m)
mu_capm <- mu_estim_capm(rets_m, ff_5factors$Mkt.RF, ff_5factors$RF)
head(mu_capm)
}
