% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_mu_estim.R
\name{mu_estim_fm}
\alias{mu_estim_fm}
\title{Factor-Model Expected Returns}
\usage{
mu_estim_fm(rets, rf, factors)
}
\arguments{
\item{rets}{an nxp data matrix.}

\item{rf}{a numeric vector or double for the risk-free return.}

\item{factors}{an nxf data frame of factors, e.g. the Fama-French 5 Research Factors.}
}
\value{
a numeric vector with the estimated expected returns.
}
\description{
Estimates the expected returns with a factor model.
The factors can be, for example, the Fama-French 5 Research Factors.
}
\details{
The expected returns are calculated according to a factor model, estimated with an OLS regression.
For a set of factors F, the expected returns for an \eqn{n\times p} data matrix X are defined as
\deqn{\hat{\mu_vec}= \hat{B}'E[F],}
where \eqn{E[F]=\frac{1}{n}\sum_{i=1}^{n}F_{i}} and \eqn{\hat{B}} are the estimated
beta coefficients from the linear regression.
}
\examples{
data(prices_m)
data(ff_5factors)
rets_m <- calc_rets(prices_m)
mu_fm <- mu_estim_fm(rets_m, ff_5factors$RF, ff_5factors[, -ncol(ff_5factors)])
head(mu_fm)
}
