% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_mu_estim.R
\name{mu_estim_ml}
\alias{mu_estim_ml}
\title{Maximum-Likelihood Expected Returns}
\usage{
mu_estim_ml(rets)
}
\arguments{
\item{rets}{an nxp matrix of the stock returns.}
}
\value{
a numeric vector with the estimated expected returns.
}
\description{
Estimates the expected returns with the Maximum-Likelihood (ML) estimator.
}
\details{
The ML estimator for the expected returns of an \eqn{n\times p} matrix with stock returns \eqn{R}
is calculated with the following formula:
\deqn{E[R]=\frac{1}{n}\sum_{i=1}^{n}R_{i}}.
}
\examples{
data(prices_m)
rets_m <- calc_rets(prices_m)
mu_ml <- mu_estim_ml(rets_m)
head(mu_ml)
}
