% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_mu_estim.R
\name{mu_estim_shrink}
\alias{mu_estim_shrink}
\title{Linear-Shrinkage Expected Returns}
\usage{
mu_estim_shrink(rets)
}
\arguments{
\item{rets}{an nxp matrix, the stock returns. n stands for the observation length, p for the number of stocks.}
}
\value{
a numeric vector with the estimated expected returns.
}
\description{
Estimates the expected returns with a linear shrinkage estimator towards the
expected return of a global minimum-variance (GMV) portfolio.
}
\details{
The Linear-Shrinkage estimator for the expected return vector of an \eqn{n\times p} matrix
with stock returns \eqn{R} is
calculated with the following formula:
\deqn{E[R]= \left(1-\delta\right)\widehat{\mu_vec} + \delta\widehat{\mu_vec}_{g}{1}_{p},}
where \eqn{\widehat{\mu_vec}} is the sample mean vector,
\eqn{\widehat{\mu_vec}_{g}} is the shrinkage target (here, the grand mean), \eqn{{1}_{p}} is a 1xp vector of ones,
and \eqn{\delta} is the optimal shrinkage intensity as in \insertCite{Stein_1956;textual}{portoptim}.
}
\examples{
data(prices_m)
rets_m <- calc_rets(prices_m)
mu_shrink <- mu_estim_shrink(rets_m)
head(mu_shrink)
}
\references{
\insertAllCited
}
