% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_mu_estim.R
\name{mu_estim_wrapper}
\alias{mu_estim_wrapper}
\title{Wrapper Function for Expected Returns Estimation I}
\usage{
mu_estim_wrapper(rets, est_func, ...)
}
\arguments{
\item{rets}{an nxp stock returns matrix.}

\item{est_func}{an estimation function.}

\item{...}{additional parameters, parsed to est_func.}
}
\value{
a numeric vector with the estimated expected returns.
}
\description{
Estimates the expected returns of a dataset
according to the user-defined function.
}
\examples{
data(prices_m)
data(ff_5factors)
rets_m <- calc_rets(prices_m)
mu_shrink <- mu_estim_wrapper(rets_m, mu_estim_shrink)
mu_fm <- mu_estim_wrapper(rets_m, mu_estim_fm, ff_5factors$RF, ff_5factors[, -ncol(ff_5factors)])

}
