% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_mu_estim.R
\name{mu_sim}
\alias{mu_sim}
\title{Simulated Expected Returns}
\usage{
mu_sim(p, min_exp_val = 0, max_exp_val = 0.5)
}
\arguments{
\item{p}{an integer, indicating the number of assets.}

\item{min_exp_val}{a double, indicating the minimum possible expected value.
Default value is 0.00.}

\item{max_exp_val}{a double, indicating the maximum possible expected value.
Default value is 0.50.}
}
\value{
a numeric vector with the simulated expected returns.
}
\description{
Simulates a vector of expected returns with specific properties.
}
\details{
The expected values are drown from an uniform distribution.
}
\examples{
mu_vec <- mu_sim(p = 100, min_exp_val = 0.00, max_exp_val = 0.50)

}
