% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_optim_trad.R
\name{port_optim_markowitz}
\alias{port_optim_markowitz}
\title{Markowitz Portfolio Optimization}
\usage{
port_optim_markowitz(sigma_mat, mu_vec, rf = 0, gamma = 2)
}
\arguments{
\item{sigma_mat}{a pxp covariance matrix of asset returns.}

\item{mu_vec}{a numeric vector of length p, the expected returns.}

\item{rf}{a double, the assumed risk-free return. Default value is 0.}

\item{gamma}{an integer, the risk aversion parameter. Default value is 2.}
}
\value{
a numeric vector with the Markowitz portfolio weights.
}
\description{
Calculates the weights of an optimal Markowitz portfolio strategy.
}
\examples{
data(prices_m)
rets_m <- calc_rets(prices_m)
sigma_mat <- cov(rets_m)
mu_vec <- mu_estim_wrapper(rets_m, mu_estim_ml)
port_markowitz <- port_optim_markowitz(sigma_mat, mu_vec)

}
