% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_optim_unb.r
\name{port_optim_markowitz_unb}
\alias{port_optim_markowitz_unb}
\title{Markowitz Portfolio Optimization (Unbiased)}
\usage{
port_optim_markowitz_unb(sigma_mat, mu_vec, rf = 0, gamma = 2, is_size)
}
\arguments{
\item{sigma_mat}{a pxp covariance matrix of asset returns.}

\item{mu_vec}{a numeric vector of length p, the expected returns.}

\item{rf}{a double, the assumed risk-free return. Default value is 0.}

\item{gamma}{an integer, the risk aversion parameter. Default value is 2.}

\item{is_size}{an integer, the number of observations of the returns data.}
}
\value{
a numeric vector with the Markowitz unbiased portfolio weights.
}
\description{
Calculates the weights of an optimal unbiased Markowitz portfolio strategy.
}
\examples{
data(prices_m)
rets_m <- calc_rets(prices_m)
n <- nrow(rets_m)
sigma_mat <- cov(rets_m)
mu_vec <- mu_estim_wrapper(rets_m, mu_estim_ml)
port_markowitz_unb <- port_optim_markowitz_unb(sigma_mat, mu_vec, is_size = n)

}
