% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_optim_trad.R
\name{port_optim_naive}
\alias{port_optim_naive}
\title{Naive Portfolio Diversification I}
\usage{
port_optim_naive(sigma_mat)
}
\arguments{
\item{sigma_mat}{a pxp covariance matrix of asset returns.}
}
\value{
a numeric vector with the naive portfolio weights.
}
\description{
Calculates the weights of a naive portfolio strategy.
}
\examples{
data(prices_m)
rets_m <- calc_rets(prices_m)
sigma_mat <- cov(rets_m)
port_naive <- port_optim_naive(sigma_mat)

}
