% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_optim_solver.R
\name{port_optim_solver}
\alias{port_optim_solver}
\title{Numerical Portfolio Optimization with Constraints}
\usage{
port_optim_solver(
  sigma_mat,
  mu_vec = NULL,
  rf = 0,
  gamma = 2,
  Aeq = NULL,
  beq = NULL,
  A = NULL,
  b = NULL,
  lambda1 = 0,
  penidx1 = NULL,
  lambda2 = 0,
  penidx2 = NULL,
  gross_c = NULL,
  porttype = "GMV",
  zero_tol = 1e-07,
  res_all = FALSE
)
}
\arguments{
\item{sigma_mat}{a pxp matrix, the covariance matrix of asset returns.}

\item{mu_vec}{a vector of length p, the expected returns.}

\item{rf}{a double, the assumed risk-free return. Default value is 0.}

\item{gamma}{an integer, the assumed risk-averse parameter. Default value is 2.}

\item{Aeq}{a cxp equality constraint matrix, containing c constraints for p regressors.
Default value is Aeq=NULL, no equality constraints.}

\item{beq}{a cx1 equality constraint vector. Default value is beq=NULL, no equality constraints.}

\item{A}{a cxp inequality constraint matrix, containing c constraints for p regressors.
Default value is A=NULL, no inequality constraints.}

\item{b}{a cx1 inequality constraint vector. Default value is b=NULL, no inequality constraints.}

\item{lambda1}{a tuning parameter value for the lasso penalty. Default value is lambda1=0.}

\item{penidx1}{a logical px1 vector,
indicating which coefficients are to be penalized with the lasso penalty lambda1.
Default value is penidx1=NULL and imposes penalty on all p coefficients.}

\item{lambda2}{a tuning parameter value for the ridge penalty. Default value is lambda2=0.}

\item{penidx2}{a logical px1 vector,
indicating which coefficients are to be penalized with the ridge penalty lambda2.
Default value is penidx2=NULL and imposes penalty on all p coefficients.}

\item{gross_c}{a double, the required gross exposure constraint.
Default value is NULL (no constraint). Attention! Works only with GMV.}

\item{porttype}{a character string.
Possible values are "Tang" (tangency portfolio) and "GMV" (global minimum variance portfolio).
Default value is "GMV".}

\item{zero_tol}{a double, indicating the zero tolerance for the calculated weights. Default value is 1e-7.}

\item{res_all}{a logical.
If TRUE, the result includes the calculated weights and the duals from the optimization.
If FALSE, only the weights. Default value is FALSE.}
}
\value{
a numeric vector with the portfolio weights.

a list with the portfolio weights and the duals from the optimization.
}
\description{
Calculates the weights of a GMV or tangency portfolio strategy with different constraints and/or penalties.
}
\details{
The portfolio optimization with constraints minimizes
\deqn{0.5||y - X \beta ||^2_2 + \lambda_1||\beta||_1 + \lambda_2||\beta||^2_2,}
subject to \eqn{Aeq \beta = beq} and \eqn{A \beta\le b}.
}
\examples{
data(prices_m)
rets_m <- calc_rets(prices_m)
sigma_mat <- cov(rets_m)
mu_vec <- mu_estim_wrapper(rets_m, mu_estim_ml)
port_gmv <- port_optim_gmv(sigma_mat)
port_gmv_solver <- port_optim_solver(sigma_mat, mu_vec, porttype = "GMV")
port_tang <- port_optim_tang(sigma_mat, mu_vec)
port_tang_solver <- port_optim_solver(sigma_mat, mu_vec, porttype = "Tang")
}
