% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_optim_trad.R
\name{port_optim_tang}
\alias{port_optim_tang}
\title{Tangency Portfolio Optimization}
\usage{
port_optim_tang(sigma_mat, mu_vec, rf = 0)
}
\arguments{
\item{sigma_mat}{a pxp matrix, the covariance matrix of asset returns.}

\item{mu_vec}{a numeric vector of length p, the expected returns.}

\item{rf}{a double, the assumed risk-free return. Default value is 0.}
}
\value{
a numeric vector with the Tangency portfolio weights.
}
\description{
Calculates the weights of a tangency portfolio strategy.
}
\examples{
data(prices_m)
rets_m <- calc_rets(prices_m)
sigma_mat <- cov(rets_m)
mu_vec <- mu_estim_wrapper(rets_m, mu_estim_ml)
port_tang <- port_optim_tang(sigma_mat, mu_vec)

}
