% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_optim.R
\name{port_optim_wrapper}
\alias{port_optim_wrapper}
\title{Wrapper Function for Portfolio Optimization I}
\usage{
port_optim_wrapper(sigma_mat, estim_func, ...)
}
\arguments{
\item{sigma_mat}{a pxp covariance matrix of asset returns.}

\item{estim_func}{a function for portfolio optimization.}

\item{...}{additional arguments to be passed to estim_func}
}
\value{
the value of the executed function estim_func.
}
\description{
Allows the execution of all included optimization functions.
}
\examples{
data(prices_m)
rets_m <- calc_rets(prices_m)
sigma_mat <- cov(rets_m)
mu_vec <- mu_estim_wrapper(rets_m, mu_estim_ml)
port_gmv <- port_optim_wrapper(sigma_mat, port_optim_gmv)
port_tang <- port_optim_wrapper(sigma_mat, port_optim_tang, mu_vec)

}
