% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prices_d.R
\docType{data}
\name{prices_d}
\alias{prices_d}
\title{200 daily stock prices}
\format{
A data frame (4547x200) with 200 daily stock prices.
}
\usage{
data(prices_d)
}
\description{
The dataset contains the daily stock prices of 200 randomly sampled stocks from the S&P 500 index
for the time period from 12.2000 to 12.2018
}
\keyword{datasets}
