% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prices_m.R
\docType{data}
\name{prices_m}
\alias{prices_m}
\title{200 monthly stock prices}
\format{
A data frame (217x200) with 200 monthly stock prices.
}
\usage{
data(prices_m)
}
\description{
The dataset contains the monthly stock prices of 200 randomly sampled stocks from the S&P 500 index
for the time period from 12.2000 to 12.2018
}
\keyword{datasets}
