% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prices_spx_d.R
\docType{data}
\name{prices_spx_d}
\alias{prices_spx_d}
\title{200 daily S&P 500 index prices}
\format{
A data frame (4547x1) with daily S&P 500 index prices.
}
\usage{
data(prices_spx_d)
}
\description{
The dataset contains the daily prices of the S&P 500 index
for the time period from 12.2000 to 12.2018
}
\keyword{datasets}
